Sidestep CI snapshotting issues: rename 'stories' to 'storybook'
