+++ components/x-privacy-manager/storybook/story-consent-accepted.js
