+++ components/x-privacy-manager/storybook/story-consent-indeterminate.js
